# exports/openapi.v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/servers/{serverName}/versions": {
      "get": {
        "description": "按站点名称列出历史版本，便于对比与回滚。",
        "operationId": "VersionsController_list",
        "parameters": [
          {
            "name": "serverName",
            "required": true,
            "in": "path",
            "description": "站点名称（nginx server_name）",
            "schema": {
              "example": "example1.com",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "版本列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServerPolicyVersionResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查看站点策略版本列表",
        "tags": [
          "ServerPolicy"
        ]
      }
    },
    "/servers/{serverName}/versions/{versionNo}/rollback": {
      "post": {
        "description": "复制目标版本生成新版本，并记录审计日志。",
        "operationId": "VersionsController_rollback",
        "parameters": [
          {
            "name": "serverName",
            "required": true,
            "in": "path",
            "description": "站点名称（nginx server_name）",
            "schema": {
              "example": "example1.com",
              "type": "string"
            }
          },
          {
            "name": "versionNo",
            "required": true,
            "in": "path",
            "description": "目标版本号",
            "schema": {
              "example": 2,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackPolicyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "回滚后的新版本",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerPolicyVersionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "回滚到指定策略版本（生成新版本）",
        "tags": [
          "ServerPolicy"
        ]
      }
    },
...
  },
  "info": {
    "title": "WAF 控制面 API",
    "description": "v1 接口文档",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ServerPolicyVersionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "版本 ID",
            "example": 12
          },
          "serverName": {
            "type": "string",
            "description": "站点名称",
            "example": "example.com"
          },
          "versionNo": {
            "type": "number",
            "description": "版本号",
            "example": 3
          },
          "policyJson": {
            "type": "object",
            "description": "策略 JSON（包含 meta/extends/rules 等）",
            "additionalProperties": true
          },
          "enabledCoreRules": {
            "description": "启用的 core 规则集名称",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "enabledTemplates": {
            "description": "启用的模板名称",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "description": "状态",
            "enum": [
              "SUCCESS",
              "FAILED",
              "ROLLED_BACK"
            ]
          },
          "note": {
            "type": "string",
            "description": "备注"
          },
          "publishLog": {
            "type": "string",
            "description": "发布日志摘要"
          },
          "createdBy": {
            "type": "string",
            "description": "创建人",
            "example": "admin"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间（ISO）"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间（ISO）"
          }
        },
        "required": [
          "id",
          "serverName",
          "versionNo",
          "policyJson",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "RollbackPolicyDto": {
        "type": "object",
        "properties": {
          "actor": {
            "type": "string",
            "description": "操作者",
            "example": "alice"
          }
        }
      },
...
    }
  }
}